refactor: 미니게임 목록 업데이트 로직 RoomService로 통합
- MiniGameSelectEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- MiniGameSelectConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 및 중복 코드 제거
